Update entityevent.ts
Since this event does not need to be cancelable, I think it should fire afterwards. This led to several hard to trace bugs in my code relating to executing commands selecting the player in this event. 

I think the event should represent after the player is joined and ready to be used. However, this is more of a design desision, so since @Rjlintkh wrote it I'm requesting them to review this and choose whether to do it.